pyproject.toml
[tool.poetry]
name = "django_template"
version = "0.1.0"
description = "Tamplate for django project"
authors = ["AlexanderZug <alexandertsygankov.unterwegs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.1"
django-grappelli = "^3.0.8"
mimesis = "^14.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
